Move the wasm-tools CLI to the `src/bin` directory (#1144)
* Move the wasm-tools CLI to the `src/bin` directory

This was a standalone `cli.rs` from when it was originally a separate
executable, but nowadays each crate has its own library interface and
all executables are slurped up into the `src/bin` directory

* Fix dependencies for just-wasm-compose
