+++ Cargo.toml
@@ -92,7 +92,7 @@ wasm-encoder = { workspace = true, optional = true }
 regex = { version = "1.6.0", optional = true }
 
 # Dependencies of `compose`
-wasm-compose = { workspace = true, optional = true, features = ['cli'] }
+wasm-compose = { workspace = true, optional = true }
 
 # Dependencies of `demangle`
 rustc-demangle = { version = "0.1.21", optional = true }
@@ -163,7 +163,7 @@ mutate = ['wasm-mutate']
 dump = ['dep:wasmparser']
 objdump = ['dep:wasmparser']
 strip = ['wasm-encoder', 'dep:wasmparser', 'regex']
-compose = ['wasm-compose']
+compose = ['wasm-compose', 'dep:wasmparser']
 demangle = ['rustc-demangle', 'cpp_demangle', 'dep:wasmparser', 'wasm-encoder']
 component = [
   'wit-component',
