+++ crates/wasm-compose/Cargo.toml
@@ -20,13 +20,8 @@ serde = { workspace = true }
 petgraph = "0.6.2"
 log = { workspace = true }
 serde_yaml = "0.9.22"
-clap = { workspace = true, optional = true }
 smallvec = "1.10.0"
 heck = "0.4.0"
-
-[features]
-default = []
-cli = ["clap"]
 
 [dev-dependencies]
 glob = "0.3.0"
