+++ crates/wasm-compose/src/lib.rs
@@ -2,8 +2,6 @@
 
 #![deny(missing_docs)]
 
-#[cfg(feature = "cli")]
-pub mod cli;
 pub mod composer;
 pub mod config;
 pub(crate) mod encoding;
